Add support for multiple SQS channels
This change adds a --sqs-profiles option to the broker CLI to start
multiple SQS channels. For each profile, credentials corresponding to
that profile will be read from the default credential file located at
~/.aws/credentials.

If no profiles are given, the default profile is read from the same
credential file path. For more information on how to configure this
file, see:
https://docs.aws.amazon.com/sdk-for-go/v2/developer-guide/configure-gosdk.html